% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gauss_fit_info_to_xslx.R
\name{add_gauss_fit_info_to_xslx}
\alias{add_gauss_fit_info_to_xslx}
\title{add_gauss_fit_info_to_xslx}
\usage{
add_gauss_fit_info_to_xslx(
  excel_file,
  excel_sheet_name,
  excel_nrows = 100,
  excel_ncols = 180,
  activate_Plots = FALSE,
  select_last_10000 = TRUE,
  alpha_packet_size = 100,
  activate_save_back_to_xlsx = TRUE
)
}
\arguments{
\item{excel_file}{Excel file name}

\item{excel_sheet_name}{Name of Excel Sheet}

\item{excel_nrows}{Number of rows to be copied out, after header (1:(excel_nrows+1)). Default: 100.}

\item{excel_ncols}{Number of columns to be copied out (1:excel_ncols). Default: 180.}

\item{activate_Plots}{if TRUE, plots are created, if FALSE, they are skipped.}

\item{select_last_10000}{If \code{TRUE} only the last 10000 Ions are selected for the Output. If \code{FALSE}
all arrival times are exported.}

\item{alpha_packet_size}{Number of alpha particles grouped into one data point, default 100.}

\item{activate_save_back_to_xlsx}{Default \code{TRUE}. If True, then the dataframe with the gauss fits is saved
back to excel. If false this step is skipped.}
}
\value{
Dataframe with gauss fit information added to it.
}
\description{
This function opens a certain sheet of an Excel file (.xlsx) and copies it out as a dataframe. It analyzes
the column "Fit Gauss Arrival Time", if this is activated (1), instead od deactivated (0), gaussian fit calculations are
performed. The stepsize is read out from the Excel Sheet, column "Fit Stepsize", if none is given a stepsize of 1000 is assumed.
The code then continues to calculate the arrival time data using \code{\link{get_data_arrival_time}}. As a default the last 10'000
ions are analyzed. If less data is available the code will analyze fewer curves and save the number of curves used into the excel file.
 In the end the fit parameters obtained with \code{\link{create_gaussian_fits}} and results are saved back into the xlsx file.
}
\details{
Excel sheet must contain a sheet named "Info", where in cell B1 the
  GEM type of the measurements is given and in cell B2 th apth to the
  folder with measurement data is given. In the general measurement data folder,
  subfolders for each GEM should be present with the GEM name as folder name.

  Folder structure:
  \itemize{
  \item{.../Measurement_Data}
    \itemize{
    \item{em1mm}
    \item{em1p5mm}
    \item{em2mm}
    \item{keramik1p5mm}
  }}

  If alph_start and alph_end in the Excel file are not yet defined (TBD),
  the default values cover the entire range of the measurement (1:End).
}
\examples{
\dontrun{
add_gauss_fit_info_to_xslx(excel_file = "em2mm_Data_Analysis.xlsx",
                           gem_type = "em2mm",
                           excel_sheet_name = "M2",
                           excel_nrows = 100,
                           activate_Plots = FALSE,
                           select_last_10000 = TRUE)

}
}
