% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_darkRate.R
\name{data_analysis_darkRate}
\alias{data_analysis_darkRate}
\title{data_analysis_darkRate}
\usage{
data_analysis_darkRate(excel_file, excel_data_frame, alpha_packet_size = 100)
}
\arguments{
\item{excel_file}{Excel File Name as a string, example: "Data_Analysis.xlsx" If the file is
in a different folder use the datapath, example: "C:/Users/username/Documents/R_Coding/Excel_File.xlsx"}

\item{excel_data_frame}{Dataframe of excel sheet to be analysed, obtained with \code{\link{extract_dataFrame_from_xlsx}}.}

\item{alpha_packet_size}{Number of alpha particles to be grouped together into one data point,
default 100.}
}
\value{
Dataframe with dark rate analysis completed.
}
\description{
Extracts the dark rate measurement data from a dataframe and then analyses it and saves the results
into the correct excel columns. Only executed if Data$Data.analysis[i]==1.
}
\examples{
\dontrun{
DataFromExcel <- extract_dataFrame_from_xlsx(excel_file = "Data_Analysis_Mobility.xlsx",
                   excel_sheet_name = "M1",
                   excel_nrows = 100,
                   excel_ncols = 180)
DataFromExcel <- data_analysis_filename(data = DataFromExcel)
DataFromExcel <- data_analysis_main(data = DataFromExcel,
                                    excel_file = "Data_Analysis_Mobility.xlsx")
}
}
