% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis_filename.R
\name{data_analysis_filename}
\alias{data_analysis_filename}
\title{data_analysis_filename}
\usage{
data_analysis_filename(excel_data_frame)
}
\arguments{
\item{excel_data_frame}{Dataframe of excel sheet to be analysed, obtained with \code{\link{extract_dataFrame_from_xlsx}}.}
}
\value{
Dataframe with file name extraction completed
}
\description{
Extracts the filename from a dataframe and then splits it at "_" and pastes the different
parameters into the correct excel columns. Only executed if Data$Filename.Read.in[i]==1.
}
\examples{
\dontrun{
DataFromExcel <- extract_dataFrame_from_xlsx(excel_file = "Data_Analysis_Mobility.xlsx",
                   excel_sheet_name = "M1",
                   excel_nrows = 100,
                   excel_ncols = 180)
DataFromExcel_with_filenames <- data_analysis_filename(data = DataFromExcel)
}
}
