% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_arrival_time.R
\name{get_data_arrival_time}
\alias{get_data_arrival_time}
\title{get_data_arrival_time}
\usage{
get_data_arrival_time(
  i = 1,
  excel_file,
  excel_sheet_name,
  excel_nrows = 100,
  excel_ncols = 180,
  select_last_10000 = FALSE,
  select_region = NA,
  alpha_packet_size = 100
)
}
\arguments{
\item{i}{Describes which row of the Excel File shall be analysed. Header does not count as row, so i=1 is the first entry.}

\item{excel_file}{Excel file name}

\item{excel_sheet_name}{Name of Excel Sheet}

\item{excel_nrows}{Number of rows to be copied out, after header (1:(excel_nrows+1)). Default: 100.}

\item{excel_ncols}{Number of columns to be copied out (1:excel_ncols). Default: 180.}

\item{select_last_10000}{If \code{TRUE} only the last 10000 Ions are selected for the Output. If \code{FALSE}
all arrival times are exported.}

\item{select_region}{c(Alpha_Start, Alpha_End). Vector of region of interest for extraction of data arrival}

\item{alpha_packet_size}{Number of alpha particles grouped into one data point, default 100.}
}
\value{
vector with arrival times in ms.
}
\description{
This function opens a certain sheet of an Excel file (.xlsx) and analyzes its \code{i}-th row
and saves the arrival time data of its ions into a vector. The arrival times are given in ms and there is an
option for only selecting the last 10'000 Ions for the calculation. Additionally, a special region of interest
can be given by using the parameter \code{select_region}.
}
\details{
Excel sheet must contain a sheet named "Info", where in cell B1 the
  GEM type of the measurements is given and in cell B2 th apth to the
  folder with measurement data is given. In the general measurement data folder,
  subfolders for each GEM should be present with the GEM name as folder name.

  Folder structure:
  \itemize{
  \item{.../Measurement_Data}
    \itemize{
    \item{em1mm}
    \item{em1p5mm}
    \item{em2mm}
    \item{keramik1p5mm}
  }}

  If alph_start and alph_end in the Excel file are not yet defined (TBD),
  the default values cover the entire range of the measurement (1:End).
}
\examples{
\dontrun{
data_arrival_time <- get_data_arrival_time(i=1,
                             excel_file = "em1p5mm_Data_Analysis.xlsx",
                             excel_sheet_name = "M1",
                             select_last_10000 = FALSE
                             )
#Arrival time of first ion:
cat("Arrival time of first ion: ", data_arrival_time[1])
}
}
